Update isnan, isinf, and isfinite.

+++ runtime/runtime.cpp
@@ -167,9 +167,9 @@ long long thorin_get_kernel_time() {
 int32_t thorin_isinff(float x)    { return std::isinf(x); }
 int32_t thorin_isnanf(float x)    { return std::isnan(x); }
 int32_t thorin_isfinitef(float x) { return std::isfinite(x); }
-int64_t thorin_isinf(double x)    { return std::isinf(x); }
-int64_t thorin_isnan(double x)    { return std::isnan(x); }
-int64_t thorin_isfinite(double x) { return std::isfinite(x); }
+int32_t thorin_isinf(double x)    { return std::isinf(x); }
+int32_t thorin_isnan(double x)    { return std::isnan(x); }
+int32_t thorin_isfinite(double x) { return std::isfinite(x); }
 
 void thorin_print_char(char c)      { std::cout << c; }
 void thorin_print_short(int16_t s)  { std::cout << s; }
